fixed-the-go-back-button-bug-in-the-passenger-details-component-and-refactored-the-baggage-files
